refactor(server): parseArgs関数を共通化し、関数名を修正
- parseCliArgs関数をconfig.tsに追加して重複コードを削除
- getStoragePathFunc()をgetStoragePath()に名前変更
