+++ packages/server/src/cli/generate-embeddings.ts
@@ -17,25 +17,9 @@ import {
   syncEmbeddingsToVectorDb,
   type GenerateEmbeddingDeps,
 } from '@/application/embedding/generate-embedding.js';
-import { initConfig, parseConfigArg } from '@/config.js';
+import { initConfig, parseCliArgs } from '@/config.js';
 import { connectDatabase, disconnectDatabase } from '@/infra/database/prisma.js';
 import { buildAppContainer } from '@/infra/di/index.js';
-
-/**
- * Parse CLI arguments, extracting command and filtering out --config option.
- */
-function parseArgs(args: string[]): { command: string; configPath: string | undefined } {
-  const configPath = parseConfigArg(args);
-  // Filter out --config and its value from args to get the command
-  const filteredArgs = args.slice(2).filter((arg, i, arr) => {
-    if (arg === '--config') return false;
-    if (i > 0 && arr[i - 1] === '--config') return false;
-    if (arg.startsWith('--config=')) return false;
-    return true;
-  });
-  const command = filteredArgs[0] ?? 'generate';
-  return { command, configPath };
-}
 
 function getDeps(): GenerateEmbeddingDeps {
   const container = buildAppContainer();
@@ -48,7 +32,7 @@ function getDeps(): GenerateEmbeddingDeps {
 }
 
 async function main(): Promise<void> {
-  const { command, configPath } = parseArgs(process.argv);
+  const { command, configPath } = parseCliArgs(process.argv);
 
   // Initialize configuration
   initConfig(configPath);
